+++ NEWS.md
@@ -6,6 +6,9 @@ likelihood when conducting signature presence test in `SigPresenceAssignActivity
 
 * Made the tracing message more informative for function `SigPresenceAssignActivity`.
 
+* Fixed a bug in function `DropLowMutationSamples` when the input spectra is not
+SBS96, SBS192, DBS78 or ID.
+
 # 2.2.1.9004
 * Fixed a bug in `AddSigActivity1`.